refactor Upload examples in different files
